requirements.txt
fastapi
requests
streamlit
uvicorn
python-dotenv
google-genai
